Perform lookups using customm gid_model_id when it is defined on a model
Previous to this commit, it was possible for a class to define a custom
model attribute to use when generating a GID. E.g.

```
class SiameseFighting < ActiveRecord::Base
  gid_model_id :name
end
```

However, the attribute defined by `::gid_model_id` was not used when
using the GID to lookup a model instance.

This commit fixes this bug by using `find_by(<gid_model_id_attribute>:
<model_id>)` when performing a lookup in
`::GlobalIdUtils::Locator.locate`. In addition to this core change, this
commit removes the ability to pass a callable object into `::gid_model_id`.
The callable pattern increased the complexity of the lookup code and is
unused by consumers of this gem (Portal).