Fix bugs in `GlobalIdUtils::Locator.locate_many`
~~ Expected behaviour ~~

When a user passes a set of GIDs into
`GlobalIdsUtils::Locator.locate_many` they expect to receive all records
which match the provided GIDs.

Furthermore, if a user does not specify the `ignore_missing` param when
calling `#locate_many`, the expectation is that an
`ActiveRecord::RecordNotFound` exception is raised if any records cannot
be found given the provided GIDs.

~~ Bug ~~

In 149ba02, we shifted the logic in `::locate_many` to use `find_by`
instead of `find`. This allowed support for lookups by other attributes
than the model's primary key. This was an attempt to fix a separate bug
wherein `::locate_many` did not respect a custom gid_model_id attribute
defined by the model class.

However, in making this change, we removed two critical behaviours of
`::locate_many`. First, `find_by` will only return one record when many
are expected. Second, `find_by` will not raise an error if a record
cannot be found.

~~ Fix ~~

This commit shifts the logic in `::locate_many` to use a `where` clause
to select records which match the provided GIDs. Additionally, logic is
added to raise an `ActiveRecord::RecordNotFound` exception if a record
could not be found for all of the provided GIDs.